Mark interface as internal
