+++ even-more-fish-plugin/src/main/java/com/oheers/fish/gui/guis/BaitsGui.java
@@ -8,7 +8,8 @@ import com.oheers.fish.config.GuiConfig;
 import com.oheers.fish.gui.ConfigGui;
 import com.oheers.fish.messages.ConfigMessage;
 import com.oheers.fish.utils.CooldownHelper;
-import com.oheers.fish.utils.Sortable;
+import com.oheers.fish.utils.sort.SortType;
+import com.oheers.fish.utils.sort.Sortable;
 import de.themoep.inventorygui.DynamicGuiElement;
 import de.themoep.inventorygui.GuiElementGroup;
 import de.themoep.inventorygui.StaticGuiElement;
@@ -28,7 +29,7 @@ public class BaitsGui extends ConfigGui {
 
     private final CooldownHelper confirmation = CooldownHelper.create();
     private final CooldownHelper cooldown = CooldownHelper.create();
-    private final Sortable.SortType sortType;
+    private final SortType sortType;
 
     public BaitsGui(@NotNull HumanEntity player) {
         super(
@@ -41,7 +42,7 @@ public class BaitsGui extends ConfigGui {
         Section config = getGuiConfig();
 
         // TODO fetch sort type from config after baits have weight.
-        this.sortType = Sortable.SortType.ALPHABETICAL;
+        this.sortType = SortType.ALPHABETICAL;
         if (config != null) {
             getGui().addElements(getBaitsGroup(config));
         }
