+++ even-more-fish-plugin/src/main/java/com/oheers/fish/utils/sort/Sortable.java
@@ -1,0 +1,13 @@
+package com.oheers.fish.utils.sort;
+
+import org.jetbrains.annotations.ApiStatus;
+import org.jetbrains.annotations.NotNull;
+
+@ApiStatus.Internal
+public interface Sortable {
+
+    double getWeight();
+
+    @NotNull String getId();
+
+}
